// web/src/lib/api/generated/models/exploreGroupDimension.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export type ExploreGroupDimension =
  (typeof ExploreGroupDimension)[keyof typeof ExploreGroupDimension];

export const ExploreGroupDimension = {
  source: "source",
  participant: "participant",
  domain: "domain",
  message_type: "message_type",
  mailing_list: "mailing_list",
  kind: "kind",
  year: "year",
  month: "month",
} as const;